Navigator.OpenFile and direct ITrayHostServices implementation
Per review feedback: Navigator should not know about 'XmlSnippets'. Drop
OpenXmlSnippets and expose a generic OpenFile(path); make Navigator
implement ITrayHostServices directly and delete the HostServices shim
classes.

- WPF/Avalonia Navigator: OpenXmlSnippets -> public OpenFile(path), and
  the class now implements ITrayHostServices.
- Avalonia: the previous private static OpenFile is promoted to the
  public instance method (same implementation, shell-execute).
- ITrayMenuContributor.Contribute now takes an ITrayHostServices argument
  so hosts pass the live Navigator on every menu-open. Plugins no longer
  [Import] ITrayHostServices, so composition has no timing dependency on
  Navigator existing before Compose().
- XmlPlugin moves the XML template-bootstrap logic (file doesn't exist
  -> write default snippet) from Navigator into its own Contribute path;
  that's XML-specific behavior and belongs with the XML plugin.
- PluginHost.AddExportedValue removed (unused after the refactor).
- Delete HostServices.cs from both host projects.
